Two changes:
1. I fixed a bug in dia processing, which I discovered today when I first
started to use dia:  the prerequisite of the dia rule was $(SRC_DIA) which is a single string conposed of filename of dia source as tokens.  When the string is used, the first file is produced number of tokens times, since make cannot access the individual tokens/filenames.

2. I have been annoyed that clean removes the generated figures.  Since I was
in the file anyway, I made removing the generated figures the highest level of
clearning.