Ported over gnuplot support.
This augments the FIGURES dependency to produce tex, png, jpg and gif
files out of gplot files.  Since it's necessary to read the gplot files to
know the target extension, intermediate .gnuplot-output files are used to
track dependencies to help rake know whether the figure needs to be
rebuilt or not.

Any data files read by a gplot file should be named with the extension
'gdata'.  All gplot files implicitly depend on all gdata files.  This
decision was made for the sake of implementation simplicity and runtime
efficiency (i.e. to avoid reading the gplot files when determining task
dependencies).