Support for Figures (pdfs) generated via R.
This follows a similar pattern to support for gnuplot, except that instead
of there being an assumed 1:1 between scripts and figures, there is an
assumed 1:1 between data and figures, and only one script.

The path to this script must be specified in the local rakefile by
assigning a value to R_CREATE_GRAPHS.  Such a path is relative to the root
of the paper (i.e. latex_dir/..)

We reuse the technique of using a .output file to help rake decide when to
regenerate the figures.

Some refactoring was done to the gnuplot rules to re-use some common code.
Additionally, a minor cleanups to the gnuplot rules was added, namely
replacing the use of
	sh "touch #{path}"
with
	FileUtils.touch path