Modify how we handle cites.
Made a large-ish change to how we decide if we should call bibtex.

1. Get list of cites in paper from .aux file.
2. Get list of available cite keys from .bib files that
   are included.
3. Get list of cite keys included in last bibtex from .bbl.

We want running bibtex to essentially be stateless.  Rerun bibtex iff:
  a) any .bib file used by master is newer than master.bbl (or
     there is a .bib and master.bbl doesn't exist)
  b) there is a cite in master.aux that is not in master.bbl but
     *is* in a bib file included by master.aux

If any bib file that the master uses is newer than the master.bbl
file then we need to run bibtex, period.  In addition, run bibtex
if there is a cite in the paper that is in one of the bib files, but
is not in the bbl file.  This avoids always running bibtex when there
are cites in the paper, but not in any .bib file, but *will* run bibtex
if there is a cite in the paper and the .bib file that is not yet
in the .bbl file.