The script fix-all-trees.rb will run git filter-branch and remove the sensitive information by running fix-one-tree.rb
Of course, these cleanup scripts have sensitive information as well, but
fortunately this happy story ends with a murder-suicide.

We should run this script (and then annihilate the modified version of
this commit) before pushing to github.

Since this does tons of damage to the repo, we should ensure that we're
fully merged before doing this, and that there are no outstanding WIP
branches.