Fixed log parsing for rerunning pdflatex
Prior to this change, every log file would be examined
for the "Rerun latex..." string.  This caused problems
when there were multiple masters since the string existing
in one log file would cause *all* masters to be built
up to MAX_ITERATION times.  Now only the log file for the
current master is examined to strings indicating the latex
should be rerun.